// src/modules/search/model/Search.ts
export interface Search {
    id: string
    image: string
    title?: string
    description?: string
    creator?: string
    view?: number
    name?: string
    level?: string
    duration?: number
    follower?: number
    rating: number
    type: string
    currentPrice?: number
    originalPrice?: number
    enroll?: number
    ratingCount?: number
    course?: number
    badges?: string[]
}